submitted-startup-pid2
To: libc-alpha@sourceware.org
Subject: [PATCH 1/2] Hurd: make reboot() send messages to both pid 1 and 2
Date: Thu, 15 Aug 2013 10:43:34 +0200

Previously /hurd/init was run as pid 1 on Hurd systems, however
sysvinit assumes that it is being run as pid 1. To appease sysvinit,
/hurd/init will be run as pid 2 in the near future.

Currently the pid of /hurd/init is hardcoded to 1. This patch makes
reboot() try to send the start_reboot message to pid 1, and if that
fails it will send it to pid 2.

* sysdeps/mach/hurd/reboot.c (reboot): Try to send a startup_reboot
  message to both pid 1 and 2.

Gbp-Pq: Topic hurd-i386
Gbp-Pq: Name submitted-startup-pid2.diff